refactor: extract routing and scheduled logic from worker/index.ts
Split worker/index.ts (551 lines) into three focused modules:
- index.ts (67 lines): thin entry point with config validation
- router.ts (357 lines): all route matching and handler dispatch
- scheduled.ts (122 lines): cron job handling

All 84 tests pass, type check clean.
