+++ worker/index.ts
@@ -1,80 +1,9 @@
-import { executePipeline } from "./state-machine";
-import { notify } from "./notify";
 import { setGitHubToken, initGitHubCircuitBreaker } from "./lib/github/index";
-import type { Env, SubmitDealBody } from "./types";
-import {
-  handleHealth,
-  handleReady,
-  handleLive,
-  handleMetrics,
-  handleGetDeals,
-  handleDiscover,
-  handleStatus,
-  handleGetLogs,
-  handleSubmit,
-  handleRankedDeals,
-  handleDealHighlights,
-  handleSimilarDeals,
-  handleExplainDeal,
-  handleAnalytics,
-} from "./routes/core";
-import {
-  handleGetReferrals,
-  handleCreateReferral,
-  handleGetReferralByCode,
-  handleDeactivateReferral,
-  handleReactivateReferral,
-} from "./routes/referrals";
-import {
-  handleResearch,
-  handleGetResearchResults,
-} from "./routes/referral-research";
+import type { Env } from "./types";
 import { jsonResponse } from "./routes/utils";
-import {
-  handleMCPRequest,
-  handleMCPListTools,
-  handleMCPCall,
-  handleMCPInfo,
-} from "./routes/mcp";
-import {
-  handleValidateUrl,
-  handleValidateBatch,
-  handleGetValidationStats,
-  handleValidateDeal,
-} from "./routes/validation";
-import {
-  handleCreateApiKey,
-  handleListApiKeys,
-  handleRevokeApiKey,
-} from "./routes/admin/keys";
-import { withAuth } from "./lib/auth";
-import { checkDealExpirations, runFullValidationSweep } from "./lib/expiration";
-import {
-  checkRateLimit,
-  getClientIdentifier,
-  createRateLimitMiddleware,
-} from "./lib/rate-limit";
-import { handleD1Request } from "./routes/d1";
-import { handleNLQRequest } from "./routes/nlq/index";
-import { handleWebhookRoutes } from "./routes/webhooks";
-import { handleSemanticSearch } from "./routes/semantic-search";
-import { logger } from "./lib/global-logger";
-import {
-  handleSubmitExperience,
-  handleGetExperience,
-  handleRunAggregation,
-} from "./routes/experience";
-import { runAggregation } from "./lib/d1/experience";
-import {
-  handleEmailIncoming,
-  handleEmailParse,
-  handleEmailHelp,
-} from "./routes/email";
 import { validateConfig } from "./lib/config-utils";
-
-// ============================================================================
-// Module-level config validation promise (prevents race condition on cold start)
-// ============================================================================
+import { handleRequest } from "./router";
+import { handleScheduled } from "./scheduled";
 
 let configValidationPromise: Promise<void> | null = null;
 
@@ -94,13 +23,8 @@ async function ensureConfigValidated(env: Env): Promise<void> {
   }
 }
 
-// ============================================================================
-// Main Worker Entry Point
-// ============================================================================
-
 export default {
   async fetch(request: Request, env: Env): Promise<Response> {
-    // Validate configuration at startup to fail fast on misconfiguration
     try {
       await ensureConfigValidated(env);
     } catch (error) {
@@ -115,319 +39,20 @@ export default {
       );
     }
 
-    // Initialize GitHub token and circuit breaker if available
     if (env.GITHUB_TOKEN) {
       setGitHubToken(env.GITHUB_TOKEN);
       initGitHubCircuitBreaker(env as unknown as { DEALS_PROD: KVNamespace });
     }
 
-    const url = new URL(request.url);
-    const path = url.pathname;
-
-    try {
-      // Health checks
-      if (path === "/health") return handleHealth(env, request);
-      if (path === "/health/ready") return handleReady(env, request);
-      if (path === "/health/live") return handleLive(env, request);
-
-      // Metrics
-      if (path === "/metrics") {
-        return withAuth(request, env, "admin", () => {
-          const format = url.searchParams.get("format") || "prometheus";
-          return handleMetrics(env, format, request);
-        });
-      }
-
-      // Deals
-      if (path === "/deals" || path === "/deals.json") {
-        return withAuth(request, env, undefined, () =>
-          handleGetDeals(url, env, request),
-        );
-      }
-      if (path === "/deals/ranked") {
-        return withAuth(request, env, undefined, () =>
-          handleRankedDeals(url, env),
-        );
-      }
-      if (path === "/deals/highlights") {
-        return withAuth(request, env, undefined, () =>
-          handleDealHighlights(url, env),
-        );
-      }
-      if (path === "/deals/similar") {
-        return withAuth(request, env, undefined, () =>
-          handleSimilarDeals(url, env),
-        );
-      }
-
-      // Pipeline API
-      if (path === "/api/discover" && request.method === "POST") {
-        return withAuth(request, env, "admin", (auth) => {
-          const rateLimiter = createRateLimitMiddleware(
-            env,
-            "/api/discover",
-            auth,
-          );
-          return rateLimiter(request, () => handleDiscover(env, request));
-        });
-      }
-      if (path === "/api/status") {
-        return withAuth(request, env, "admin", () =>
-          handleStatus(env, request),
-        );
-      }
-      if (path === "/api/log") {
-        return withAuth(request, env, "admin", () =>
-          handleGetLogs(url, env, request),
-        );
-      }
-      if (path === "/api/analytics") {
-        return withAuth(request, env, "admin", () =>
-          handleAnalytics(url, env, request),
-        );
-      }
-
-      // Deal Submission
-      if (path === "/api/submit" && request.method === "POST") {
-        return withAuth(request, env, "user", (auth) => {
-          const rateLimiter = createRateLimitMiddleware(
-            env,
-            "/api/submit",
-            auth,
-          );
-          return rateLimiter(request, () => handleSubmit(request, env));
-        });
-      }
-
-      // Referral Action Routes (deactivate/reactivate) - Moved up to avoid shadowing
-      const referralActionMatch = path.match(
-        /^\/api\/referrals\/([^/]+)\/(deactivate|reactivate)$/,
-      );
-      if (referralActionMatch && request.method === "POST") {
-        const code = referralActionMatch[1];
-        const action = referralActionMatch[2];
-
-        if (code && action === "deactivate") {
-          return withAuth(request, env, "user", () =>
-            handleDeactivateReferral(request, code, env),
-          );
-        }
-        if (code && action === "reactivate") {
-          return withAuth(request, env, "user", () =>
-            handleReactivateReferral(code, env),
-          );
-        }
-      }
-
-      // Referral API
-      if (path === "/api/referrals") {
-        if (request.method === "GET") {
-          return withAuth(request, env, undefined, () =>
-            handleGetReferrals(url, env),
-          );
-        }
-        if (request.method === "POST") {
-          return withAuth(request, env, "user", () =>
-            handleCreateReferral(request, env),
-          );
-        }
-      }
-
-      // Referral Detail Route (GET by code)
-      const referralDetailMatch = path.match(/^\/api\/referrals\/([^/]+)$/);
-      if (referralDetailMatch && request.method === "GET") {
-        const code = referralDetailMatch[1];
-        if (code) {
-          return withAuth(request, env, undefined, () =>
-            handleGetReferralByCode(code, env, request),
-          );
-        }
-      }
-
-      // Research API
-      if (path === "/api/research" && request.method === "POST") {
-        return withAuth(request, env, "user", (auth) => {
-          const rateLimiter = createRateLimitMiddleware(
-            env,
-            "/api/research",
-            auth,
-          );
-          return rateLimiter(request, () => handleResearch(request, env));
-        });
-      }
-
-      // Research results API
-      if (path.startsWith("/api/research/") && request.method === "GET") {
-        const domain = path.replace("/api/research/", "");
-        return withAuth(request, env, undefined, () =>
-          handleGetResearchResults(domain, env),
-        );
-      }
-
-      // Semantic search (Vectorize + Workers AI)
-      if (path === "/api/semantic-search" && request.method === "POST") {
-        return withAuth(request, env, "user", () =>
-          handleSemanticSearch(request, env),
-        );
-      }
-
-      // Validation API
-      if (path === "/api/validate/url" && request.method === "POST") {
-        return withAuth(request, env, "user", (auth) =>
-          handleValidateUrl(request, env, auth),
-        );
-      }
-      if (path === "/api/validate/batch" && request.method === "POST") {
-        return withAuth(request, env, "user", (auth) =>
-          handleValidateBatch(request, env, auth),
-        );
-      }
-      if (path === "/api/validation/stats" && request.method === "GET") {
-        return withAuth(request, env, "admin", () =>
-          handleGetValidationStats(env, request),
-        );
-      }
-
-      const dealExplainMatch = path.match(/^\/api\/deals\/([^/]+)\/explain$/);
-      if (dealExplainMatch && request.method === "GET") {
-        const dealId = dealExplainMatch[1] ?? "";
-        return withAuth(request, env, undefined, () =>
-          handleExplainDeal(dealId, env, request),
-        );
-      }
-
-      const dealValidateMatch = path.match(/^\/api\/deals\/([^/]+)\/validate$/);
-      if (dealValidateMatch && request.method === "POST") {
-        const code = dealValidateMatch[1] ?? "";
-        return withAuth(request, env, "user", (auth) =>
-          handleValidateDeal(request, code, env, auth),
-        );
-      }
-
-      // MCP (Model Context Protocol) Endpoints - 2025-11-25 Specification
-      if (path === "/mcp") {
-        return withAuth(request, env, "user", () =>
-          handleMCPRequest(request, env),
-        );
-      }
-
-      // Legacy MCP v1 Endpoints (for backwards compatibility)
-      if (path === "/mcp/v1/tools/list" && request.method === "POST") {
-        return withAuth(request, env, "user", () =>
-          handleMCPListTools(env, request),
-        );
-      }
-      if (path === "/mcp/v1/tools/call" && request.method === "POST") {
-        return withAuth(request, env, "user", () =>
-          handleMCPCall(request, env),
-        );
-      }
-      if (path === "/mcp/v1/info") {
-        return withAuth(request, env, "user", () =>
-          handleMCPInfo(env, request),
-        );
-      }
-
-      // D1 Database API endpoints
-      if (path.startsWith("/api/d1/")) {
-        return withAuth(request, env, "admin", () =>
-          handleD1Request(request, url, env),
-        );
-      }
-
-      // NLQ (Natural Language Query) API endpoints
-      if (path.startsWith("/api/nlq")) {
-        return withAuth(request, env, "user", () =>
-          handleNLQRequest(request, url, env),
-        );
-      }
-
-      // Webhook routes
-      const webhookPath = path.startsWith("/api") ? path.slice(4) : path;
-      const webhookResponse = await handleWebhookRoutes(
-        request,
-        env,
-        webhookPath,
-      );
-      if (webhookResponse) return webhookResponse;
-
-      // Experience Feedback API
-      if (path === "/api/experience" && request.method === "POST") {
-        return withAuth(request, env, "user", () =>
-          handleSubmitExperience(request, env),
-        );
-      }
-
-      const experienceMatch = path.match(/^\/api\/experience\/([^/]+)$/);
-      if (experienceMatch && request.method === "GET") {
-        if (experienceMatch[1] !== undefined)
-          return withAuth(request, env, undefined, () =>
-            handleGetExperience(experienceMatch[1]!, env),
-          );
-      }
-
-      if (path === "/api/experience/aggregate" && request.method === "POST") {
-        return withAuth(request, env, "admin", () =>
-          handleRunAggregation(env, request),
-        );
-      }
-
-      // Email API
-      if (path === "/api/email/incoming" && request.method === "POST") {
-        return handleEmailIncoming(request, env);
-      }
-      if (path === "/api/email/parse" && request.method === "POST") {
-        return withAuth(request, env, "user", () =>
-          handleEmailParse(request, env),
-        );
-      }
-      if (path === "/api/email/help" && request.method === "GET") {
-        return withAuth(request, env, undefined, () => handleEmailHelp());
-      }
-
-      // Admin API Key Management
-      if (path === "/api/admin/keys") {
-        if (request.method === "POST") {
-          return withAuth(request, env, "admin", () =>
-            handleCreateApiKey(request, env),
-          );
-        }
-        if (request.method === "GET") {
-          return withAuth(request, env, "admin", () =>
-            handleListApiKeys(request, env),
-          );
-        }
-      }
-
-      const apiKeyRevokeMatch = path.match(/^\/api\/admin\/keys\/([^/]+)$/);
-      if (apiKeyRevokeMatch && request.method === "DELETE") {
-        const hash = apiKeyRevokeMatch[1];
-        if (hash) {
-          return withAuth(request, env, "admin", () =>
-            handleRevokeApiKey(request, hash, env),
-          );
-        }
-      }
-
-      // 404
-      return jsonResponse({ error: "Not found" }, 404, request, env);
-    } catch (error) {
-      console.error("Request handler error:", error);
-      return jsonResponse(
-        { error: "Internal server error", message: (error as Error).message },
-        500,
-        request,
-        env,
-      );
-    }
+    return handleRequest(request, env);
   },
 
   async scheduled(event: ScheduledEvent, env: Env): Promise<void> {
-    // Validate configuration at startup to fail fast on misconfiguration
     try {
       validateConfig(env);
     } catch (error) {
       console.error("Scheduled execution configuration error:", error);
+      const { notify } = await import("./notify");
       await notify(env, {
         type: "system_error",
         severity: "critical",
@@ -437,115 +62,6 @@ export default {
       return;
     }
 
-    const cron = event.cron;
-    const timestamp = new Date().toISOString();
-
-    logger.info(`Scheduled event triggered: ${cron}`, {
-      component: "scheduled",
-      cron,
-      timestamp,
-    });
-
-    try {
-      // Daily cron job at 9am - expiration checks and experience aggregation
-      if (cron === "0 9 * * *") {
-        logger.info("Running daily expiration check", {
-          component: "scheduled",
-        });
-
-        const result = await checkDealExpirations(env);
-
-        logger.info("Daily expiration check completed", {
-          component: "scheduled",
-          expiringFound: result.expiringFound,
-          expiredMarked: result.expiredMarked,
-          notificationsSent: result.notificationsSent,
-        });
-
-        logger.info("Running daily experience aggregation", {
-          component: "scheduled",
-        });
-
-        const aggResult = await runAggregation(env.DEALS_DB!);
-
-        logger.info("Daily experience aggregation completed", {
-          component: "scheduled",
-          dealsProcessed: aggResult.dealsProcessed,
-          eventsProcessed: aggResult.eventsProcessed,
-        });
-
-        return;
-      }
-
-      // Weekly cron job - full validation sweep
-      if (cron === "0 0 * * 0") {
-        logger.info("Running weekly full validation sweep", {
-          component: "scheduled",
-        });
-
-        const result = await runFullValidationSweep(env);
-
-        logger.info("Weekly validation sweep completed", {
-          component: "scheduled",
-          validated: result.validated,
-          deactivated: result.deactivated,
-          expiringNotified: result.expiringNotified,
-          errors: result.errors.length,
-        });
-
-        if (result.errors.length > 0) {
-          await notify(env, {
-            type: "system_error",
-            severity: "warning",
-            run_id: `weekly-validation-${Date.now()}`,
-            message: `Weekly validation completed with ${result.errors.length} errors`,
-            context: {
-              errors: result.errors,
-              validated: result.validated,
-              deactivated: result.deactivated,
-            },
-          });
-        }
-
-        return;
-      }
-
-      // Default pipeline execution (every 6 hours)
-      logger.info("Running pipeline execution", {
-        component: "scheduled",
-      });
-
-      const result = await executePipeline(env);
-      if (!result.success) {
-        console.error(`Pipeline failed at ${result.phase}: ${result.error}`);
-        await notify(env, {
-          type: "system_error",
-          severity: "critical",
-          run_id: `pipeline-${Date.now()}`,
-          message: `Pipeline failed at ${result.phase}: ${result.error}`,
-          context: {
-            phase: result.phase,
-            error: result.error,
-          },
-        });
-      } else {
-        logger.info("Pipeline execution completed successfully", {
-          component: "scheduled",
-          phase: result.phase,
-        });
-      }
-    } catch (error) {
-      console.error("Scheduled execution error:", error);
-      await notify(env, {
-        type: "system_error",
-        severity: "critical",
-        run_id: "scheduled",
-        message: `Scheduled execution failed: ${(error as Error).message}`,
-        context: {
-          cron,
-          error: (error as Error).message,
-        },
-      });
-    }
+    return handleScheduled(event, env);
   },
 };
